Almost there
The "2 columns" need to be moved to the bottom of the page, as well as have the "h3 tags" show up above the lists instead of next to them.
Also need to add pictures (logo next to the name & pictures in the middle).
Finally, if there's time, add a bootswatch theme.